Move server_default tests in a separate file
